Add plaforms/worker entry
